+++ src/metagen/framework/domain/literals.py
@@ -30,9 +30,11 @@ D_META = Literal["DYNAMIC"]
 D: Final = "DYNAMIC"
 S_META = Literal["STATIC"]
 S: Final = "STATIC"
+P_META = Literal["PERMUTATION"]
+P: Final = "PERMUTATION"
 
 METAGEN_TYPE = Literal["DEFINITION", "INTEGER",
-                       "REAL", "CATEGORICAL", "DYNAMIC", "STATIC"]
+                       "REAL", "CATEGORICAL", "DYNAMIC", "STATIC", "PERMUTATION"]
 
 BacicVal: TypeAlias = Union[int, float, str, List[int], List[float], List[str]]
 DefVal: TypeAlias = Mapping[str, Union[BacicVal, "DefVal"]]
@@ -56,6 +58,7 @@ DefAttr: TypeAlias = Tuple[DF_META, Mapping[str, Union[BaseAttr, "DefAttr"]]]
 StrBaseAttr: TypeAlias = Union["Attributes", Tuple["Attributes", ...]]
 DymAttr: TypeAlias = Tuple[D_META, int, int, Optional[int], Optional[StrBaseAttr]]
 StaAttr: TypeAlias = Tuple[S_META, int, Optional[StrBaseAttr]]
-Attributes: TypeAlias = Union[BaseAttr, DefAttr, DymAttr, StaAttr]
+PermAttr: TypeAlias = Tuple[P_META, CatVal]
+Attributes: TypeAlias = Union[BaseAttr, DefAttr, DymAttr, StaAttr, PermAttr]
 
 DefType: TypeAlias = Mapping[str, Attributes]